tömbök folytatása

+++ 06_Tömbök/Solution_Tömbök/Feladat_07_gyümölcs/Fruit.cs
@@ -1,7 +1,7 @@
 ﻿public class Fruit
 {
     public string Name { get; set; }
-    public int Amount { get; set; }
+    public double Amount { get; set; }
 
     public double UnitPrice { get; set; }
 
@@ -12,7 +12,7 @@
 
     }
 
-    public Fruit(string name, int amount, double unitPrice)
+    public Fruit(string name, double amount, double unitPrice)
     {
         Name = name;
         Amount = amount;
